Enabled Python tests. Fixed some bindings
