+++ .gitlab-ci.yml
@@ -4,6 +4,7 @@ variables:
   pacman_deps_makepkg: rsync fakeroot git namcap
   pacman_deps_py: python python-pip
   pacman_deps_test: gtest
+  pacman_deps_pytest: python-pytest python-numpy
 
 stages:
   - build
@@ -42,6 +43,7 @@ test_py:
     - pacman -Syu --noconfirm --needed ${pacman_deps} ${pacman_deps_py} ${pacman_deps_test}
     - chmod a+rX -R .
     - su nobody -s /bin/bash -c "BUILD_WITH_PYTHON=1 BUILD_WITH_TESTING=1 ./scripts/build.sh"
+    # todo: make package available and run python tests
 
 build_archlinux:
   stage: build
@@ -71,7 +73,7 @@ test_archlinux_py:
   image: ${image}
   needs: [build_archlinux]
   script:
-    - pacman -Syu --noconfirm --needed ${pacman_deps} ${pacman_deps_py} ${pacman_deps_test}
+    - pacman -Syu --noconfirm --needed ${pacman_deps} ${pacman_deps_py} ${pacman_deps_test} ${pacman_deps_pytest}
     - pacman -U --noconfirm ./structstore{,_py}-*.pkg.tar.zst
     - chmod a+rX -R .
     - su nobody -s /bin/bash -c "BUILD_WITH_PYTHON=1 ./scripts/test.sh"
@@ -95,4 +97,4 @@ test_pywheel:
   script:
     - pacman -Syu --noconfirm --needed ${pacman_deps} ${pacman_deps_py} ${pacman_deps_test}
     - chmod a+rX -R .
-    - su nobody -s /bin/bash -c "BUILD_WITH_PYTHON=1 ./scripts/test.sh"
+    - su nobody -s /bin/bash -c "BUILD_WITH_PYTHON=1 BUILD_WITH_VENV=1 ./scripts/test.sh"
