+++ pyproject.toml
@@ -65,3 +65,4 @@ nanobind = "^2.1.0"
 nanobind-stubgen = "^0.1.5"
 numpy = "<2"
 build = "^1.2.2"
+pytest = "^8.3.3"
