+++ requirements.txt
@@ -4,18 +4,24 @@ build==1.2.2 ; python_version >= "3.9" and python_version < "4.0" \
 click==8.1.7 ; python_version >= "3.9" and python_version < "4.0" \
     --hash=sha256:ae74fb96c20a0277a1d615f1e4d73c8414f5a98db8b799a7931d1582f3390c28 \
     --hash=sha256:ca9853ad459e787e2192211578cc907e7594e294c7ccc834310722b41b9ca6de
-colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and (os_name == "nt" or platform_system == "Windows") \
+colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and (os_name == "nt" or sys_platform == "win32" or platform_system == "Windows") \
     --hash=sha256:08695f5cb7ed6e0531a20572697297273c47b8cae5a63ffc6d6ed5c201be6e44 \
     --hash=sha256:4f1d9991f5acc0ca119f9d443620b77f9d6b33703e51011c16baf57afb285fc6
 distlib==0.3.8 ; python_version >= "3.9" and python_version < "4.0" \
     --hash=sha256:034db59a0b96f8ca18035f36290806a9a6e6bd9d1ff91e45a7f172eb17e51784 \
     --hash=sha256:1530ea13e350031b6312d8580ddb6b27a104275a31106523b8f123787f494f64
+exceptiongroup==1.2.2 ; python_version >= "3.9" and python_version < "3.11" \
+    --hash=sha256:3111b9d131c238bec2f8f516e123e14ba243563fb135d3fe885990585aa7795b \
+    --hash=sha256:47c2edf7c6738fafb49fd34290706d1a1a2f4d1c6df275526b62cbb4aa5393cc
 flit-core==3.9.0 ; python_version >= "3.9" and python_version < "4.0" \
     --hash=sha256:72ad266176c4a3fcfab5f2930d76896059851240570ce9a98733b658cb786eba \
     --hash=sha256:7aada352fb0c7f5538c4fafeddf314d3a6a92ee8e2b1de70482329e42de70301
 importlib-metadata==8.5.0 ; python_version >= "3.9" and python_full_version < "3.10.2" \
     --hash=sha256:45e54197d28b7a7f1559e60b95e7c567032b602131fbd588f1497f47880aa68b \
     --hash=sha256:71522656f0abace1d072b9e5481a48f07c138e00f079c38c8f883823f9c26bd7
+iniconfig==2.0.0 ; python_version >= "3.9" and python_version < "4.0" \
+    --hash=sha256:2d91e135bf72d31a410b17c16da610a82cb55f6b0477d1a902134b24a455b8b3 \
+    --hash=sha256:b6a85871a79d2e3b22d2d1b94ac2824226a63c6b741c88f7ae975f18b6778374
 nanobind-stubgen==0.1.5 ; python_version >= "3.9" and python_version < "4.0" \
     --hash=sha256:3f4e3c852dfb6ce2bd51fa5e710f3019180176669280fd00e7293e08dee76709
 nanobind==2.1.0 ; python_version >= "3.9" and python_version < "4.0" \
@@ -60,12 +66,18 @@ numpy==1.26.4 ; python_version >= "3.9" and python_version < "4.0" \
 packaging==24.1 ; python_version >= "3.9" and python_version < "4.0" \
     --hash=sha256:026ed72c8ed3fcce5bf8950572258698927fd1dbda10a5e981cdf0ac37f4f002 \
     --hash=sha256:5b8f2217dbdbd2f7f384c41c628544e6d52f2d0f53c6d0c3ea61aa5d1d7ff124
+pluggy==1.5.0 ; python_version >= "3.9" and python_version < "4.0" \
+    --hash=sha256:2cffa88e94fdc978c4c574f15f9e59b7f4201d439195c3715ca9e2486f1d0cf1 \
+    --hash=sha256:44e1ad92c8ca002de6377e165f3e0f1be63266ab4d554740532335b9d75ea669
 py-build-cmake==0.1.8 ; python_version >= "3.9" and python_version < "4.0" \
     --hash=sha256:bee4107849d2646df7c1b461e1268e24109dd903e57d14bc99d32864a04e802d \
     --hash=sha256:f90b8e2b9a273670fe7251e744eaede75132e798e5402477282a47012b2e5bdb
 pyproject-hooks==1.1.0 ; python_version >= "3.9" and python_version < "4.0" \
     --hash=sha256:4b37730834edbd6bd37f26ece6b44802fb1c1ee2ece0e54ddff8bfc06db86965 \
     --hash=sha256:7ceeefe9aec63a1064c18d939bdc3adf2d8aa1988a510afec15151578b232aa2
+pytest==8.3.3 ; python_version >= "3.9" and python_version < "4.0" \
+    --hash=sha256:70b98107bd648308a7952b06e6ca9a50bc660be218d53c257cc1fc94fda10181 \
+    --hash=sha256:a6853c7375b2663155079443d2e45de913a911a11d669df02a50814944db57b2
 tomli==2.0.1 ; python_version >= "3.9" and python_version < "3.11" \
     --hash=sha256:939de3e7a6161af0c887ef91b7d41a53e7c5a1ca976325f429cb46ea9bc30ecc \
     --hash=sha256:de526c12914f0c550d15924c62d72abc48d6fe7364aa87328337a31007fe8a4f
