+++ scripts/build_and_test.sh
@@ -1,5 +1,22 @@
 #!/bin/bash -ex
 
-BUILD_WITH_TESTING=1 ./scripts/build.sh
-cd /tmp/structstore_build
-ctest --output-on-failure
+srcdir="$PWD"
+pkgdir="/tmp/structstore_build"
+
+BUILD_WITH_TESTING=1 "$srcdir"/scripts/build.sh
+ctest --output-on-failure "$pkgdir"
+
+if [ ! -z "$BUILD_WITH_PYTHON" ]; then
+    # build wheel
+    rm -rf "$pkgdir"
+    "$srcdir"/scripts/build_py.sh
+    test -f "$srcdir"/dist/structstore-*.whl
+
+    # install wheel
+    venvdir="$pkgdir/venv"
+    python -m venv "$venvdir"
+    source "$venvdir/bin/activate"
+    pip install -r "$srcdir/requirements.txt"
+    pip install "$srcdir"/dist/structstore-*.whl
+    pytest "$srcdir/tests"
+fi
